code refactoring

+++ app/src/main/java/com/example/mobileterminalsproject/MainActivity.kt
@@ -241,7 +241,7 @@ class MainActivity : AppCompatActivity(){
         mapResponseYT = Gson().fromJson(jsonObjectYT.toString(), mapResponseYT.javaClass)
         //extracting all video ids and adding them to a list
         val items = mapResponseYT["items"] as ArrayList<*>
-        for(i in 0 until textButtonRadio.toInt())
+        for(i in 0 until textButtonRadio)
             videoIdList.add((((items[i] as LinkedTreeMap<*, *>)["id"] as LinkedTreeMap<*, *>)["videoId"]).toString())
     }
 }
